Fix openssl in krm-functions images
Users behind firewalls may be unable to build all the airshipctl
images as a result of OpenSSL errors stemming from certificates used
with the proxy. To address this, the images impacted have been
updated to copy the proxy certificate from airshipctl/certs into
the image and run update-ca-certificates command. This is the same
behavior that already exists in the base airshipctl image and is
simply being extended to the images in krm-functions.

Change-Id: I436d1e25fb0726ab7576fec44a7ef2f9fd20f2bd
Signed-off-by: Alexander Hughes <Alexander.Hughes@pm.me>