Add airship-in-a-pod to gate jobs
This change aims to add the airship-in-a-pod test to the gates.
This is accomplished by creating new scripts to run which
install a minimal k8s environment, and then apply the
airship-in-a-pod.yaml.

Brief description of scripts

01_dns_settings: Running k8s pods in zuul has a lot of issues with
DNS resolution which this resolves.

10_install_minikube: grabs latest minikube and other dependencies.

11_build_images: builds the images under tools/airship-in-a-pod
and then adds patches to the kustomize configuration to use and
configure them.

12_start_minikube: starts minikube and makes sure kubectl will work

13_apply_dns.sh: This takes some of the effort from 01_dns_settings
and incorporates it into the minikube cluster by modifying coredns

20_apply_aiap: Applies the airship-in-a-pod.yaml to the cluster.
Has a few loops to check that the cluster is fully running
before applying, and to check on the pod to see if it is ready
before finishing.

Closes: #478
Change-Id: I4c02d04b0eda9395642032bf56b56a0f823c1969