Improve release notes issue selection accuracy
The existing behavior was to select issues closed between two tag
dates. This is error prone since issues may be closed after the
fact, and there may be multiple releases being worked at the same
time and it is impossible to tell which release an issue was
closed for.

This switches to selecting issues by matching their milestone name
against the git tag. Since github does not support assigning
multiple assignments, their will likely need to be some manual
issue additions to the release notes when an issue is targeted to
multiple releases e.g. v2.0.1 and v2.1.

This also moves to only triggering github release creation on
non-pre-release semver tags, to avoid cluttering the release
list.

Change-Id: I99824020789323f8a1db418c73acef57ebb86022
Signed-off-by: Sean Eagan <seaneagan1@gmail.com>