Separate folders-Create airshipctl cli user guide documentation
Current modified go code, generates RsT files for each airshipctl
cmd and places them in a folder specific to the airshipctl subcommand.
It also generates the index.rst file for the specific airshipctl
subcommand folders. These generated rst files are appropriate to sphinx,
so that they can be easliy rendered in docs.airshipit.org

Steps to render the document locally

	* Install sphinx: python3 -m pip install sphinx
        * Clone airshipctl: git clone https://github.com/airshipit/airshipctl.git
        * Download current PS: cd airshipctl; git fetch "https://review.opendev.org/airship/airshipctl" refs/changes/50/789250/2 && git checkout -b change-789250-1 FETCH_HEAD
	* Build sphinx html pages: cd docs/source; sphinx-build -b html . _build
	* Run local server: cd _build; python3 -m http.server
	* Open URL to access the page: http://localhost:8000/ navigate to commands section access the document

Relates-To: #280
Change-Id: Ifa1dafc7b296014cc826072ac2c52f4488421352