Updating cmd files for documentation
This PS updates cmd related to root, completion, document and
version subcommands. It also updates minor details in other
subcommands.

The description and examples are updated for the airshipctl
commands, which will be inturn used for generating documentation.
Please ignore the .md file changes in this PS. They are added for zuul
gates to pass. Here is the PS with generated documention
files https://review.opendev.org/c/airship/airshipctl/+/789250

Relates-To: #280
Change-Id: I562c15c0d25b2e9731c0eb03854d1d348eb435f2