Add `replacements/` for the flux composite
The flux composite bundles together the two individual flux operator
functions, but it fails to provide a `replacements/` to bundle together
their replacements.  This results in sites in airshipctl and
treasuremap needing to pull in the function replacementes individually
at the site level, negating the benefit of the composite abstraction.

This adds the replacements kustomization and adjusts test-site to use it.

Change-Id: Ib08e24bdd6ec8222363137bd09ad79dbcd804490