Redirect all stdout of the toolbox containers to stderr
Initially we did not want this because we had a feeling that
at some point we want to perform render action inside the script.
Rendering inside the script would facilitate better debugging
functionality. Where each script would be able to render, validate
and run. However, at this point solution seems like overengineering
and seems like this should be simply a limitation of the toolbox image.

Change-Id: I99871c1c79934612d9a1ac06a5141e57e5497054