Add kubectl-wait-cluster-ephemeral phase
This adds a phase to wait for the target cluster to become initialized,
directed at the ephemeral management cluster.  It re-uses the helper script
already used by the `kubectl-wait-cluster-target` phase, adds parameterization
so that it can watch for both controlPlaneInitialized and controlPlaneReady
conditions, and adds the new phase to the phase plan.

Lack of this waiting was causing errors in some environments.

This also adds a kubectl-wait-node-target phase definition,
in case we ever want to make use of it.

Change-Id: Ic18742dd59a7901b0d985b8c11260bc1b3527a30