AIAP: support deploying manifests from patchsets
`airshipctl document pull` does not yet support pull from git refs,
and so cannot pull airshipctl patchsets to test their manifests.

As a workaround, this change copies the airshipctl project directory
that is used to build the binary / run the scripts
(which can be set to a patchset via $MANIFEST_REPO_REF) to the
expected /tmp location for serving manifests from patchsets as well.

Support for testing treasuremap manifest patchsets is still unsupported.

Change-Id: Ia9ec248efa2a12728e9eb6d3753b35ec5bae0d47