Implement fetching remote refs for document pull
This commit adds the `fetch` and `fetch.remoteRefSpec` fields to the
configuration for repositories, allowing the capability to fetch
references from a remote. This is useful for pulling specific gerrit
patchsets.

This also changes `checkout.remoteRefSpec` to `checkout.refSpec`, and
implements the logic required to checkout to an arbitrary ref.

Closes: #616

Change-Id: Ie21a6c2a7a7ac92ed3c05fef7e5683203cd62e45