AIAP: Remove `tty` setting from libvirt container
The `tty` setting may be required in some instances dependent upon the
environment in which AIAP is deployed. However, these instances are much
less frequent than successful deployments when the setting is set to
false. Further investigation will be needed to determine when it is
appropriate to use this flag.

Change-Id: Ibbd44cbf0697a71f8081cd170dbdf5f7b4e4f106