Add a script to run AIAP in AKS
This adds a script that a developer can use to run airship in a pod
in a dynamically created AKS cluster.  By default a resource group is
created and then cleaned up following the test.

Further work needs to be done to surface some indication of a failed
test out of the pod, and watch for that condition in this script.

Change-Id: I29f462b9becd6d1de0a9e0e1d7877c6019cbd6ee