Fix AIAP manifest branch checkout behavior
Even when specifying a manifest branch in the AIAP overrides, the
entrypoint script overrides/disables the checkout behavior of the
airshipctl document pull script if AIRSHIP_CONFIG_PHASE_REPO_REF is not
set. This change updates the entrypoint script to check if
AIRSHIP_CONFIG_PHASE_REPO_BRANCH is specified before overriding the
checkout behavior.

Signed-off-by: Drew Walters <drewwalters@microsoft.com>
Change-Id: Ibb2798201f8e02333289f193355f3aa2a6c627ad